Simplify to use a single Popover component
