+++ src/link/edit.js
@@ -6,11 +6,13 @@ import {
 	insert,
 	applyFormat,
 	removeFormat,
+	useAnchor,
 } from '@wordpress/rich-text';
 import { __ } from '@wordpress/i18n';
 import { InlineEditUI } from './inline';
 import { PreviewEditUI } from './preview';
 import { CustomTooltip } from './tooltip';
+import { Popover } from '@wordpress/components';
 
 const formatType = 'wikipediapreview/link';
 const formatTitle = __( 'Wikipedia Preview', 'wikipedia-preview' );
@@ -42,12 +44,30 @@ const Edit = ( {
 } ) => {
 	const [ addingPreview, setAddingPreview ] = useState( false );
 	const startAddingPreview = () => setAddingPreview( true );
+	// const startAddingPreview = () => {
+	// 	console.log('startAddingPreview');
+	// 	setAddingPreview( true );
+	// 	// if ( contentRef.current ) {
+	// 	// 	contentRef.current.focus();
+	// 	// }
+	// };
 	const stopAddingPreview = () => setAddingPreview( false );
 	const [ viewingPreview, setViewingPreview ] = useState( false );
 	const startViewingPreview = () => setViewingPreview( true );
+	// const startViewingPreview = () => {
+	// 	setViewingPreview( true );
+	// 	if ( contentRef.current ) {
+	// 		contentRef.current.focus();
+	// 	}
+	// };
 	const stopViewingPreview = () => setViewingPreview( false );
 	const [ lastValue, setLastValue ] = useState( null );
 	const toolbarButtonRef = useRef();
+	const anchor = useAnchor( {
+		editableContentElement: contentRef.current,
+		value,
+		settings,
+	} );
 
 	const formatButtonClick = () => {
 		if ( isActive ) {
@@ -100,9 +120,13 @@ const Edit = ( {
 	};
 
 	const goToEdit = () => {
+		console.log('goToEdit - value:', value);
+		console.log('goToEdit - contentRef:', contentRef);
+		console.log('goToEdit - activeAttributes:', activeAttributes);
+		
 		startAddingPreview();
 		stopViewingPreview();
-		onFocus();
+		// onFocus();
 	};
 
 	const onClosePreview = () => {
@@ -158,6 +182,7 @@ const Edit = ( {
 	};
 
 	const handleTextEdit = () => {
+		console.log('handleTextEdit');
 		// Assuming a Left-To-Right language:
 		// --> cursorDirection > 0 means cursor is moving left
 		// --> cursorDirection < 0 means cursor is moving right
@@ -174,6 +199,7 @@ const Edit = ( {
 			const formatStart = getFormatStart( value.start - 1 );
 			const formatEnd = getFormatEnd( value.end + 1 );
 			onChange( removeFormat( value, formatType, formatStart, formatEnd ) );
+			// setLastValue( null );
 		}
 	};
 
@@ -189,11 +215,16 @@ const Edit = ( {
 	useEffect( () => {
 		if ( lastValue === null ) {
 			setLastValue( value );
-		} else if ( lastValue !== value ) {
+		} else if ( lastValue.text.length !== value.text.length ) {
 			handleTextEdit();
 			setLastValue( value );
 		}
 	}, [ value ] );
+
+	useEffect( () => {
+		console.log('edit.js - adding/viewing useEffect  addingPreview:', addingPreview);
+		console.log('edit.js - adding/viewing useEffect viewingPreview:', viewingPreview);
+	}, [ addingPreview, viewingPreview ] );
 
 	return (
 		<>
@@ -212,32 +243,37 @@ const Edit = ( {
 				anchorRef={ toolbarButtonRef }
 				addingPreview={ addingPreview }
 			/>
-			{ addingPreview && (
-				<InlineEditUI
-					contentRef={ contentRef }
-					settings={ settings }
-					onApply={
-						value.start !== value.end || activeAttributes.title
-							? updateAttributes
-							: insertText
-					}
-					value={ value }
-					activeAttributes={ activeAttributes }
-					onClose={ stopAddingPreview }
-				/>
-			) }
-			{ viewingPreview && (
-				<PreviewEditUI
-					contentRef={ contentRef }
-					settings={ settings }
-					value={ value }
-					onClose={ onClosePreview }
-					onEdit={ goToEdit }
-					onRemove={ removeAttributes }
-					onForceClose={ stopViewingPreview }
-					activeAttributes={ activeAttributes }
-				/>
-			) }
+			{ ( addingPreview || viewingPreview ) && (
+				<Popover
+					anchor={ anchor }
+					placement={ addingPreview ? 'top' : 'bottom' }
+					noArrow={ false }
+					expandOnMobile={ true }
+					onClose={ addingPreview ? stopAddingPreview : onClosePreview }
+					className={`wikipediapreview-edit-${ addingPreview ? 'inline' : 'preview-popover' }`}
+					// onClick={ onClickPopoverOutside }
+				>
+					{ addingPreview && (
+						<InlineEditUI
+							onApply={
+								value.start !== value.end || activeAttributes.title
+									? updateAttributes
+									: insertText
+							}
+							value={ value }
+							activeAttributes={ activeAttributes }
+						/>
+					) }
+					{ viewingPreview && (
+						<PreviewEditUI
+							onEdit={ goToEdit }
+							onRemove={ removeAttributes }
+							onForceClose={ stopViewingPreview }
+							activeAttributes={ activeAttributes }
+						/>
+					) }
+				</Popover>
+			)}
 		</>
 	);
 };
